Implemented the suggested changes
