+++ AdvanceTaskMarsPart1/AdvanceTaskMarsPart1.csproj
@@ -20,7 +20,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="ExtentReports\" />
     <Folder Include="Screenshot\" />
   </ItemGroup>
