+++ AdvanceTaskMarsPart1/Tests/Profile_Tests.cs
@@ -1,5 +1,4 @@
 ﻿using AdvanceTaskMarsPart1.Utilities;
-using AventStack.ExtentReports;
 using NUnit.Framework;
 using AdvanceTaskMarsPart1.Steps;
 
@@ -10,7 +9,6 @@ namespace AdvanceTaskMarsPart1.Tests
     {
         LoginSteps loginSteps;
         ProfileSteps profileSteps;
-        public static ExtentTest test;
 
         public Profile_Tests()
         {
@@ -19,9 +17,8 @@ namespace AdvanceTaskMarsPart1.Tests
         }
 
         [SetUp]
-        public void SetUp()
+        public void LoginActions()
         {
-            Initialize();
             loginSteps.doLogin();
         }
 
@@ -29,41 +26,27 @@ namespace AdvanceTaskMarsPart1.Tests
         [TestCase(1)]
         public void Edit_Availability(int id)
         {
-            test = extent.CreateTest("Edit_Availability").Info("Test Started");
             profileSteps.editAvailability(id);
-            test.Log(Status.Pass, "Edit_Availability passed");
         }
         
         [Test, Order(2), Description("This test is cancelling Availability type in the Profile")]
         public void Cancel_Availability()
         {
-            test = extent.CreateTest("Cancel_Availability").Info("Test Started");
             profileSteps.cancelAvailability();
-            test.Log(Status.Pass, "Cancel_Availability passed");
         }
         
         [Test, Order(3), Description("This test is updating Hours in the Profile")]
         [TestCase(2)]
         public void Edit_Hours(int id)
         {
-            test = extent.CreateTest("Edit_Hours").Info("Test Started");
             profileSteps.editHours(id);
-            test.Log(Status.Pass, "Edit_Hours passed");
         }
 
         [Test, Order(4), Description("This test is updating EarnTarget in the Profile")]
         [TestCase(3)]
         public void Edit_EarnTarget(int id)
         {
-            test = extent.CreateTest("Edit_EarnTarget").Info("Test Started");
             profileSteps.editEarnTarget(id);
-            test.Log(Status.Pass, "Edit_EarnTarget passed");
-        }  
-
-        [TearDown]
-        public void TearDown()
-        {
-            Close();
-        }
+        } 
     }
 }
