+++ AdvanceTaskMarsPart1/Tests/SearchSkills_Tests.cs
@@ -1,5 +1,4 @@
 ﻿using AdvanceTaskMarsPart1.Utilities;
-using AventStack.ExtentReports;
 using NUnit.Framework;
 using AdvanceTaskMarsPart1.Steps;
 
@@ -10,7 +9,6 @@ namespace AdvanceTaskMarsPart1.Tests
     {
         LoginSteps loginSteps;
         SearchSkillSteps searchSkillSteps;
-        public static ExtentTest test;
 
         public SearchSkills_Tests()
         {
@@ -19,32 +17,21 @@ namespace AdvanceTaskMarsPart1.Tests
         }
 
         [SetUp]
-        public void SetUp()
+        public void LoginActions()
         {
-            Initialize();
             loginSteps.doLogin();
         }
 
         [Test, Order(1), Description("This test is returning the skills by category")]
         public void SearchSkills_Category()
         {
-            test = extent.CreateTest("SearchSkills_Category").Info("Test started");
-            searchSkillSteps.searchSkillCategory();           
-            test.Log(Status.Pass, "SearchSkills_Category passed");
+            searchSkillSteps.searchSkillCategory();
         }
 
         [Test, Order(2), Description("This test is returning the skills by filters")]
         public void SearchSkills_Filters()
         {
-            test = extent.CreateTest("SearchSkills_Filters").Info("Test started");
             searchSkillSteps.searchSkillFilters();
-            test.Log(Status.Pass, "SearchSkills_Filters passed");           
         } 
-
-        [TearDown]
-        public void TearDown()
-        {
-            Close();
-        }
     }
 }
