+++ AdvanceTaskMarsPart1/Utilities/BaseSetUp.cs
@@ -12,46 +12,55 @@ namespace AdvanceTaskMarsPart1.Utilities
     {
         public static WebDriver driver;
         public static ExtentReports extent;
+        public static ExtentTest test;
 
         [OneTimeSetUp]
         public static void ExtentStart()
         {
-            //Create new instance of ExtentReports
-            extent = new ExtentReports();
-            //Create new instance of ExtentSparkReporter
-            var SparkReporter = new ExtentSparkReporter(@"D:\Sasikala\MVP_Studio\AdvanceTaskPart1\AdvanceTaskMarsPart1\AdvanceTaskMarsPart1\ExtentReports\Index.html");
-            //Attach the ExtentSparkReporter to the ExtentReports
-            extent.AttachReporter(SparkReporter);
+            if (extent == null)
+            {
+                extent = new ExtentReports();
+                var SparkReporter = new ExtentSparkReporter("D:\\Sasikala\\MVP_Studio\\AdvanceTaskPart1\\AdvanceTaskMarsPart1\\AdvanceTaskMarsPart1\\ExtentReport\\ExtentReport.html");
+                extent.AttachReporter(SparkReporter);
+            }
         }
 
-        public void Initialize()
+        [SetUp]
+        public void SetUp()
         {
             driver = new ChromeDriver();
             driver.Manage().Window.Maximize();
             driver.Navigate().GoToUrl("http://localhost:5000/Home");
+            test = extent.CreateTest(TestContext.CurrentContext.Test.Name);
         }
 
-        public static void CaptureScreenshot(string testName)
+        [TearDown]
+        public void TearDown()
         {
-            // Capture the screenshot
-            string screenshotFileName = $"screenshot_{testName}";
-            ITakesScreenshot ts = (ITakesScreenshot)driver;
-            Screenshot screenshot = ts.GetScreenshot();
-            string filePath = "D:\\Sasikala\\MVP_Studio\\AdvanceTaskPart1\\AdvanceTaskMarsPart1\\AdvanceTaskMarsPart1\\Screenshot";
-            string screenshotPath = Path.Combine(filePath, $"{screenshotFileName}_{DateTime.Now:yyyyMMdd_HHmmss}.png");
-            screenshot.SaveAsFile(screenshotPath);
-        }
-
-        public void Close()
-        {
+            //If tests fails capture screenshot
+            if (TestContext.CurrentContext.Result.Outcome.Status == NUnit.Framework.Interfaces.TestStatus.Failed)
+            {
+                string testName = TestContext.CurrentContext.Test.Name;
+                test.Log(Status.Fail, $"Tests  '{testName}' failed");
+                CaptureScreenshot(testName);
+            }
             driver.Quit();
         }
 
         [OneTimeTearDown]
         public static void ExtentClose()
         {
-            //Flush the ExtentReports instance
             extent.Flush();
+        }
+
+        public void CaptureScreenshot(string testName)
+        {
+            string screenshotFileName = $"screenshot_{testName}";
+            ITakesScreenshot ts = (ITakesScreenshot)driver;
+            Screenshot screenshot = ts.GetScreenshot();
+            string filePath = "D:\\Sasikala\\MVP_Studio\\AdvanceTaskPart1\\AdvanceTaskMarsPart1\\AdvanceTaskMarsPart1\\Screenshot";
+            string screenshotPath = Path.Combine(filePath, $"{screenshotFileName}_{DateTime.Now:yyyyMMdd_HHmmss}.png");
+            screenshot.SaveAsFile(screenshotPath);
         }
     }
 }
